Update description to be more specific when checking if new package is needed
When it comes to the case that a package is been added to the repository, it's essential to know that where that logic is truly needed. Meaning that the package's component or function has been specifically imported to the code logic.

Feeling the term `you do need these package` might not be clear enough when it comes to the case that an automatic dependency management system helps you adding some packages due to transiting requirements by other dependencies, but it's not actually need for the service itself. Let the language being more specific will help the developers not assuming that the auto-addition does not mean it's truly needed, you might still need to head back to the logic and checkout if there is specific imports.